fix linters and update readme
