+++ Gemfile
@@ -1,0 +1,1 @@
+gem 'rubocop', '>= 1.0', '< 2.0'
