+++ README.md
@@ -1,47 +1,153 @@
-<!-- @format -->
+<a name="readme-top"></a>
 
-# Morse_Message
+<div align="center">
+  <h3><b>Enumerable</b></h3>
+</div>
 
-In this project we are creating some methods to decode morse message
+<!-- TABLE OF CONTENTS -->
 
-## Getting Started
+# 📗 Table of Contents
 
-## Built with
+- [📖 About the Project](#about-project)
+  - [🛠 Built With](#built-with)
+    - [Tech Stack](#tech-stack)
+- [💻 Getting Started](#getting-started)
+  - [Setup](#setup)
+  - [Prerequisites](#prerequisites)
+  - [Install](#install)
+  - [Usage](#usage)
+- [👥 Authors](#authors)
+- [🤝 Contributing](#contributing)
+- [⭐️ Show your support](#support)
+- [🙏 Acknowledgements](#acknowledgements)
+- [❓ FAQ](#faq)
+- [📝 License](#license)
 
-- IRB
-- `Ruby`
+<!-- PROJECT DESCRIPTION -->
 
-<p align="right">(<a href="#top">back to top</a>)</p>
+# 📖 Enumerable <a name="about-project"></a>
 
-# Authors :bookmark_tabs:
 
-👤 **Choice Osobor**
+**Enumerable** is a project that implements a subset of the Enumerable mixin. We implemented 3 methods which are all?, any? and filter.
 
-- GitHub: [@techEmprez](https://github.com/techEmprez)
-- Twitter: [@techEmprez](https://twitter.com/techEmprez)
-- LinkedIn: [Choice Osobor](https://www.linkedin.com/in/choice-osobor/)
+## 🛠 Built With <a name="built-with"></a>
+
+### Tech Stack <a name="tech-stack"></a>
+
+<details>
+  <summary>Server</summary>
+  <ul>
+    <li><a href="https://www.ruby-lang.org/en/">Ruby</a></li>
+  </ul>
+</details>
+
+<p align="right"><a href="#readme-top">back to top</a></p>
+
+
+<!-- GETTING STARTED -->
+
+## 💻 Getting Started <a name="getting-started"></a>
+
+To get a local copy up and running, follow these steps.
+
+### Prerequisites
+
+In order to run this project you need:
+
+```sh
+ gem install rails
+```
+
+### Setup
+
+Clone this repository to your desired folder:
+
+```sh
+  cd enumerable
+  git clone https://github.com/Elerqsousy/enumerable.git
+```
+
+### Install
+
+Install this project with:
+
+```sh
+  gem install
+```
+
+### Usage
+
+To run the project, execute the following command:
+
+Example command:
+
+```sh
+  ruby my_list.rb
+```
+
+<p align="right"><a href="#readme-top">back to top</a></p>
+
+<!-- AUTHORS -->
+
+## 👥 Authors <a name="authors"></a>
+
+👤 **Jeremiah Melfah**
+
+- GitHub: [GitHub](https://github.com/Jaymelfah)
+- LinkedIn: [LinkedIn](https://www.linkedin.com/in/jeremiah-ekow-melfah-a4402a161/)
+
+👤 **Ahmed Nazir Musah**
+
+- GitHub: [@ahmednazirmusah](https://github.com/AhmedNazirMusah)
+- Twitter: [@ahmednazirmush](https://twitter.com/ahmednazirmusah)
+- LinkedIn: [LinkedIn](https://www.linkedin.com/in/ahmed-nazir-musah-529956214)
 
 👤 **Mahmoud Rizk**
 
 - GitHub: [@Elerqsousy](https://github.com/Elerqsousy)
 - LinkedIn: [Mahmoud (Riz) El Erqsousy](https://www.linkedin.com/in/mahmoud-rizk-elerqsousy/)
 
-## 🤝 Contributing
+<p align="right"><a href="#readme-top">back to top</a></p>
+
+
+<!-- CONTRIBUTING -->
+
+## 🤝 Contributing <a name="contributing"></a>
 
 Contributions, issues, and feature requests are welcome!
 
 Feel free to check the [issues page](../../issues/).
 
-## Show your support
+<p align="right"><a href="#readme-top">back to top</a></p>
 
-Give a ⭐️ if you like this project!
+<!-- SUPPORT -->
 
-## Acknowledgments
+## ⭐️ Show your support <a name="support"></a>
 
-- Hat tip to anyone whose code was used
-- Inspiration
-- etc
+If you like this project...
 
-## 📝 License
+<p align="right"><a href="#readme-top">back to top</a></p>
 
-This project is [MIT](./MIT.md) licensed.
+<!-- FAQ (optional) -->
+
+## ❓ FAQ <a name="faq"></a>
+
+- **Do I have to have any previous knowledge about ruby to use this project?**
+
+  - No!
+
+- **How to use this project?**
+
+  - check the [Getting Started](#getting-started) section to have a complete guid on how to setup and run teh project.
+
+<p align="right"><a href="#readme-top">back to top</a></p>
+
+<!-- LICENSE -->
+
+## 📝 License <a name="license"></a>
+
+This project is [MIT](./LICENSE) licensed.
+
+_NOTE: we recommend using the [MIT license](https://choosealicense.com/licenses/mit/) - you can set it up quickly by [using templates available on GitHub](https://docs.github.com/en/communities/setting-up-your-project-for-healthy-contributions/adding-a-license-to-a-repository). You can also use [any other license](https://choosealicense.com/licenses/) if you wish._
+
+<p align="right"><a href="#readme-top">back to top</a></p>
