+++ my_enumerable.rb
@@ -1,0 +1,19 @@
+module MyEnumerable
+  def all?
+    result = true
+    each { |e| result = false unless yield e }
+    result
+  end
+
+  def any?
+    result = false
+    each { |e| result = true if yield e }
+    result
+  end
+
+  def filter
+    result = []
+    each { |e| result << e if yield e }
+    result
+  end
+end
